# api/swagger.yaml
definitions:
...
  v4.Pagination:
    properties:
      count:
        description: The amount of records returned in this response
        example: 25
        type: integer
      limit:
        description: The maximum amount of resources to return for this request
        example: 25
        type: integer
      offset:
        description: The offset for the first record returned
        example: 50
        type: integer
      total:
        description: The total number of resources matching the query
        example: 827
        type: integer
    type: object
...
  v4.Transaction:
    properties:
      amount:
        description: The maximum value is "999999999999.99999999", swagger unfortunately
          rounds this.
        example: 14.03
        maximum: 1000000000000
        minimum: 1e-08
        multipleOf: 1e-08
        type: number
      availableFrom:
        description: The date from which on the transaction amount is available for
          budgeting. Only used for income transactions. Defaults to the transaction
          date.
        example: "2021-11-17T00:00:00Z"
        type: string
      createdAt:
        description: Time the resource was created
        example: "2022-04-02T19:28:44.491514Z"
        type: string
      date:
        description: Date of the transaction. Time is currently only used for sorting
        example: "1815-12-10T18:43:00.271152Z"
        type: string
      deletedAt:
        description: Time the resource was marked as deleted
        example: "2022-04-22T21:01:05.058161Z"
        type: string
      destinationAccountId:
        description: ID of the destination account
        example: 8e16b456-a719-48ce-9fec-e115cfa7cbcc
        type: string
      envelopeId:
        description: ID of the envelope
        example: 2649c965-7999-4873-ae16-89d5d5fa972e
        type: string
      id:
        description: UUID for the resource
        example: 65392deb-5e92-4268-b114-297faad6cdce
        type: string
      importHash:
        description: The SHA256 hash of a unique combination of values to use in duplicate
          detection
        example: 867e3a26dc0baf73f4bff506f31a97f6c32088917e9e5cf1a5ed6f3f84a6fa70
        type: string
      links:
        $ref: '#/definitions/v4.TransactionLinks'
      note:
        description: A note
        example: Lunch
        type: string
      reconciledDestination:
        default: false
        description: Is the transaction reconciled in the destination account?
        example: true
        type: boolean
      reconciledSource:
        default: false
        description: Is the transaction reconciled in the source account?
        example: true
        type: boolean
      sourceAccountId:
        description: ID of the source account
        example: fd81dc45-a3a2-468e-a6fa-b2618f30aa45
        type: string
      updatedAt:
        description: Last time the resource was updated
        example: "2022-04-17T20:14:01.048145Z"
        type: string
    type: object
  v4.TransactionCreateResponse:
    properties:
      data:
        description: List of created Transactions
        items:
          $ref: '#/definitions/v4.TransactionResponse'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.TransactionEditable:
    properties:
      amount:
        description: The maximum value is "999999999999.99999999", swagger unfortunately
          rounds this.
        example: 14.03
        maximum: 1000000000000
        minimum: 1e-08
        multipleOf: 1e-08
        type: number
      availableFrom:
        description: The date from which on the transaction amount is available for
          budgeting. Only used for income transactions. Defaults to the transaction
          date.
        example: "2021-11-17T00:00:00Z"
        type: string
      date:
        description: Date of the transaction. Time is currently only used for sorting
        example: "1815-12-10T18:43:00.271152Z"
        type: string
      destinationAccountId:
        description: ID of the destination account
        example: 8e16b456-a719-48ce-9fec-e115cfa7cbcc
        type: string
      envelopeId:
        description: ID of the envelope
        example: 2649c965-7999-4873-ae16-89d5d5fa972e
        type: string
      importHash:
        description: The SHA256 hash of a unique combination of values to use in duplicate
          detection
        example: 867e3a26dc0baf73f4bff506f31a97f6c32088917e9e5cf1a5ed6f3f84a6fa70
        type: string
      note:
        description: A note
        example: Lunch
        type: string
      reconciledDestination:
        default: false
        description: Is the transaction reconciled in the destination account?
        example: true
        type: boolean
      reconciledSource:
        default: false
        description: Is the transaction reconciled in the source account?
        example: true
        type: boolean
      sourceAccountId:
        description: ID of the source account
        example: fd81dc45-a3a2-468e-a6fa-b2618f30aa45
        type: string
    type: object
  v4.TransactionLinks:
    properties:
      self:
        description: The transaction itself
        example: https://example.com/api/v4/transactions/d430d7c3-d14c-4712-9336-ee56965a6673
        type: string
    type: object
  v4.TransactionListResponse:
    properties:
      data:
        description: List of transactions
        items:
          $ref: '#/definitions/v4.Transaction'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/v4.Pagination'
        description: Pagination information
    type: object
...
  v4.TransactionResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v4.Transaction'
        description: The Transaction data, if creation was successful
      error:
        description: The error, if any occurred for this transaction
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.httpError:
    properties:
      error:
        example: An ID specified in the query string was not a valid UUID
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v4/transactions:
    get:
      description: Returns a list of transactions
      parameters:
      - description: Date of the transaction. Ignores exact time, matches on the day
          of the RFC3339 timestamp provided.
        in: query
        name: date
        type: string
      - description: Transactions at and after this date. Ignores exact time, matches
          on the day of the RFC3339 timestamp provided.
        in: query
        name: fromDate
        type: string
      - description: Transactions before and at this date. Ignores exact time, matches
          on the day of the RFC3339 timestamp provided.
        in: query
        name: untilDate
        type: string
      - description: Filter by amount
        in: query
        name: amount
        type: string
      - description: Amount less than or equal to this
        in: query
        name: amountLessOrEqual
        type: string
      - description: Amount more than or equal to this
        in: query
        name: amountMoreOrEqual
        type: string
      - description: Filter by note
        in: query
        name: note
        type: string
      - description: Filter by budget ID
        in: query
        name: budget
        type: string
      - description: Filter by ID of associated account, regardeless of source or
          destination
        in: query
        name: account
        type: string
      - description: Filter by source account ID
        in: query
        name: source
        type: string
      - description: Filter by destination account ID
        in: query
        name: destination
        type: string
      - description: Filter by envelope ID
        in: query
        name: envelope
        type: string
      - description: Reconcilication state in source account
        in: query
        name: reconciledSource
        type: boolean
      - description: Reconcilication state in destination account
        in: query
        name: reconciledDestination
        type: boolean
      - description: The offset of the first Transaction returned. Defaults to 0.
        in: query
        name: offset
        type: integer
      - description: Maximum number of Transactions to return. Defaults to 50.
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.TransactionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.TransactionListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.TransactionListResponse'
      summary: Get transactions
      tags:
      - Transactions
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      responses:
        "204":
          description: No Content
      summary: Allowed HTTP verbs
      tags:
      - Transactions
    post:
      description: Creates transactions from the list of submitted transaction data.
        The response code is the highest response code number that a single transaction
        creation would have caused. If it is not equal to 201, at least one transaction
        has an error.
      parameters:
      - description: Transactions
        in: body
        name: transactions
        required: true
        schema:
          items:
            $ref: '#/definitions/v4.TransactionEditable'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v4.TransactionCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.TransactionCreateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.TransactionCreateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.TransactionCreateResponse'
      summary: Create transactions
      tags:
      - Transactions
  /v4/transactions/{id}:
    delete:
      description: Deletes a transaction
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Delete transaction
      tags:
      - Transactions
    get:
      description: Returns a specific transaction
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
      summary: Get transaction
      tags:
      - Transactions
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Allowed HTTP verbs
      tags:
      - Transactions
    patch:
      consumes:
      - application/json
      description: Updates an existing transaction. Only values to be updated need
        to be specified.
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      - description: Transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/v4.TransactionEditable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.TransactionResponse'
      summary: Update transaction
      tags:
      - Transactions
...
swagger: "2.0"